[新增]1.新增修改與刪除發文內容的功能。2.新增限制搜尋範圍的功能。3.在會員中心可以看到粉絲名單。4.可以觀看其他會員中心的資料。 [修改]1.修改首頁貼文的追蹤按鈕觸發事件，當點擊時，需要更新追蹤者的搜尋列表。2.將原本的搜尋頁面改為首頁使用。3.將登入功能改為使用dialog方式顯示於首頁。4.修正定位功能的其他串接功能。

+++ static/js/model/membersetting-m.js
@@ -24,7 +24,7 @@ class memberSettingModel {
         requestAnimationFrame(() => {
             requestAnimationFrame(() => {
                 setTimeout(() => {
-                    window.location.replace("/eatsmap");
+                    window.location.replace("/");
                 }, 300);
                 
             });
@@ -38,15 +38,22 @@ class memberSettingModel {
             credentials: "include",
         });
 
-        sessionStorage.clear();
-        requestAnimationFrame(() => {
+        const dt = await response.json();
+        if (dt.logOut === true){
+            sessionStorage.clear();
+            // 清除
+            document.cookie = "my_eatweb_isLogged_here=; max-age=0; path=/;";
+            
             requestAnimationFrame(() => {
-                setTimeout(() => {
-                    window.location.replace("/login");
-                }, 300);
-                
+                requestAnimationFrame(() => {
+                    setTimeout(() => {
+                        window.location.replace("/");
+                    }, 300);
+                    
+                });
             });
-        });
+        }
+        
     }
 
     async updateMemberInfo(id) {
